NOAA ISD metadata data.frame

This data.frame includes metadata describing all the data provided in ISD data files. And is used for transforming and scaling variables.

Format

A data frame with 643 rows and 19 columns

Details

Original csv data is in inst/extdata/isd_metadata.csv, collected from

The data.frame has the following columns:

pos - (chr) position, if any

category - (chr) category, one of additional-data section, control-data section, element quality data section, mandatory-data section, original observation data section, or remarks data section

sub_category - (chr) sub category label, one of climate reference network unique data, cloud and solar data, ground surface data, hail data, marine data, network metadata, precipitation-data, pressure data, runway visual range data, sea surface temperature, soil temperature data, temperature data, weather occurrence data, weather-occurrence-data, or wind data

abbrev - (chr) abbreviation, if any, NA for control and mandatory sections

label - (chr) label, a top level label for the data, usually the same as the abbreviation

sub_label - (chr) sub label, a more detailed label about the variable

field_length - (int) field length, number of characters

min - (chr) minimum value, if applicable, original

min_numeric - (int) minimum value, if applicable, integer

max - (chr) maximum value, if applicable, original

max_numeric - (chr) maximum value, if applicable, integer

units - (chr) units, if applicable

scaling_factor - (chr) scaling factor, original

scaling_factor_numeric - (int) scaling factor, integer, one of 1, 10, 100, 1000, or NA

missing - (chr) value used to indicate missing data, original

missing_numeric - (int) value used to indicate missing data, integer, one of 9, 99, 999, 9999, 99999, 999999, or NA

description - (chr) short description of variable

dom - (chr) long description of variable with categories

dom_parsed_json - (list) NA if no categories, or a named list with category labels and their values